[fix] Fix double data type

+++ rental_app/lib/screens/account_screen.dart
@@ -42,7 +42,7 @@ class AccountScreen extends ConsumerWidget {
                 kycStatus = data['kyc']['status'],
                 avatarUrl = data['avatarUrl'] ??=
                     'https://firebasestorage.googleapis.com/v0/b/rental-app-dcdbf.appspot.com/o/app_files%2Favatar.png?alt=media&token=0b9a2456-3c04-458b-a319-83f5717c5cd4';
-            final double? balance = data['wallet']?['balance'];
+            final double? balance = data['wallet']?['balance'].toDouble();
             final bool kycVerified = data['kyc']['verified'];
             final bool hasShop = data['shop']['hasShop'];
             final bool isVerified = data['kyc']['verified'];
